Move csv-validator to csv folder
